Add deletion directly from localStorage and add callback to refresh the list instead of deleting HTML. Working flawlessly!
